Conflicts:
	Common/ac/spritecache.cpp
	Common/util/compress.cpp
	Engine/ac/draw.cpp
	Engine/ac/draw.h
	Engine/ac/dynamicsprite.cpp
	Engine/ac/game.cpp
	Engine/ac/room.cpp
	Engine/ac/sprite.cpp
	Engine/gfx/ali3dogl.cpp
	Engine/gfx/ali3dsw.cpp
	Engine/main/engine.cpp
	Engine/main/game_run.h
	Engine/platform/windows/gfx/ali3dd3d.cpp
	Solutions/Common.Lib/Common.Lib.vcproj

+++ Common/font/wfnfontrenderer.cpp
@@ -86,7 +86,7 @@ int WFNFontRenderer::GetTextWidth(const char *texx, int fontNumber)
     char* fontaddr = psp_get_char(foon, thisCharacter);
 
     short tabaddr_d;
-    memcpy(&tabaddr_d, (char*)((int)fontaddr + 0), 2);
+    memcpy(&tabaddr_d, (char*)((long)fontaddr + 0), 2);
 
     totlen += tabaddr_d;
 #endif
@@ -114,7 +114,7 @@ int WFNFontRenderer::GetTextHeight(const char *texx, int fontNumber)
 #else
     char* fontaddr = psp_get_char(foon, thisCharacter);
     short tabaddr_d;
-    memcpy(&tabaddr_d, (char*)((int)fontaddr + 2), 2);
+    memcpy(&tabaddr_d, (char*)((long)fontaddr + 2), 2);
 
     int charHeight = tabaddr_d;
 #endif
@@ -155,10 +155,10 @@ int WFNFontRenderer::printchar(int xxx, int yyy, wgtfont foo, int charr)
   char* tabaddr = psp_get_char(foo, charr);
 
   short tabaddr_d;
-  memcpy(&tabaddr_d, (char*)((int)tabaddr), 2);
+  memcpy(&tabaddr_d, (char*)((long)tabaddr), 2);
   int charWidth = tabaddr_d;
 
-  memcpy(&tabaddr_d, (char*)((int)tabaddr + 2), 2);
+  memcpy(&tabaddr_d, (char*)((long)tabaddr + 2), 2);
   int charHeight = tabaddr_d;
 #endif
