+++ Engine/ac/global_screen.cpp
@@ -12,7 +12,7 @@
 #include "ac/roomstruct.h"
 #include "ac/runtime_defines.h"
 #include "ac/screen.h"
-#include "debug/debug.h"
+#include "debug/debug_log.h"
 #include "platform/base/agsplatformdriver.h"
 #include "gfx/graphicsdriver.h"
 #include "gfx/bitmap.h"
@@ -27,7 +27,7 @@ extern roomstruct thisroom;
 extern IGraphicsDriver *gfxDriver;
 extern AGSPlatformDriver *platform;
 extern color palette[256];
-extern unsigned long loopcounter;
+extern unsigned int loopcounter;
 
 int scrnwid,scrnhit;
 int current_screen_resolution_multiplier = 1;
